% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta_feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Perform Boruta Feature Selection}
\usage{
feature_selection(traindata, target_variable, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{target_variable}{The target variable to predict.}

\item{output_file}{The file name to save the Boruta results.}
}
\description{
This function performs Boruta feature selection on the training data and saves the results.
}
