% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_baseline_table.R
\name{generate_baseline_table}
\alias{generate_baseline_table}
\title{Generate Baseline Table}
\usage{
generate_baseline_table(traindata, categorical_vars, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{categorical_vars}{The indices of categorical variables in the data.}

\item{output_file}{The file name to save the baseline table.}
}
\description{
This function generates a baseline table for categorical variables in the training data.
}
