% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_feature_selection.R
\name{lasso_feature_selection}
\alias{lasso_feature_selection}
\title{Perform LASSO Feature Selection}
\usage{
lasso_feature_selection(traindata, target_variable, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{target_variable}{The target variable to predict.}

\item{output_file}{The file name to save the LASSO results.}
}
\description{
This function performs LASSO feature selection on the training data and saves the results.
}
