% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_logistic_regression.R
\name{multivariate_logistic_regression}
\alias{multivariate_logistic_regression}
\title{Perform Multivariate Logistic Regression}
\usage{
multivariate_logistic_regression(traindata, target_variable, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{target_variable}{The target variable to predict.}

\item{output_file}{The file name to save the multivariate regression results.}
}
\description{
This function performs multivariate logistic regression on the training data and saves the results.
}
