% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_main.R
\name{selector}
\alias{selector}
\title{Bootstrap selection of most appropriate test statistic.}
\usage{
selector(f, b.reps = 1000, seed = 1)
}
\arguments{
\item{f}{Object obtained by fitting a model using sem() or cfa() in lavaan.}

\item{b.reps}{Number of bootstrap samples. Defaults to 1000.}

\item{seed}{A random seed. Defaults to 1.}
}
\value{
A plot of bootstrap p-values and the selected test statistic and its associated p-value
}
\description{
This function takes a fitted lavaan object and bootstraps various test statistics.
The test whose p-values are closest to a uniform distribution is selected.
}
