% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Data into Training and Testing Sets}
\usage{
split_data(data, target_variable, train_proportion)
}
\arguments{
\item{data}{The input data frame.}

\item{target_variable}{The target variable to predict.}

\item{train_proportion}{The proportion of data to be used for training.}
}
\value{
A list containing the training and testing sets.
}
\description{
This function splits a dataset into training and testing sets based on the target variable and train proportion.
}
