% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_regression.R
\name{stepwise_regression}
\alias{stepwise_regression}
\title{Perform Stepwise Logistic Regression}
\usage{
stepwise_regression(traindata, target_variable, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{target_variable}{The target variable to predict.}

\item{output_file}{The file name to save the stepwise regression results.}
}
\description{
This function performs stepwise logistic regression on the training data and saves the results.
}
