% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_rfe_feature_selection.R
\name{svm_rfe_feature_selection}
\alias{svm_rfe_feature_selection}
\title{Perform SVM-RFE Feature Selection}
\usage{
svm_rfe_feature_selection(traindata, target_variable, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{target_variable}{The target variable to predict.}

\item{output_file}{The file name to save the SVM-RFE results.}
}
\description{
This function performs SVM-RFE feature selection on the training data and saves the results.
}
