% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pvalues.R
\name{test_pvalues}
\alias{test_pvalues}
\title{Test for correct model specification with various test statistics.}
\usage{
test_pvalues(f)
}
\arguments{
\item{f}{A lavaan object produced by sem() or cfa()}
}
\value{
A set of pvalues associated with different test statistics
}
\description{
This function takes a lavaan fit object and calculates the p-values for various
test of correct model specification. The tests include normal-theory ml likelihood ratio test,
the satorra-bentler test and the scaled-and-shifte test. In addition less known tests
such as the scaled F test, and three tests based on eigenvalue block averaging.
}
