% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_logistic_regression.R
\name{univariate_logistic_regression}
\alias{univariate_logistic_regression}
\title{Perform Univariate Logistic Regression}
\usage{
univariate_logistic_regression(traindata, target_variable, output_file)
}
\arguments{
\item{traindata}{The training data frame.}

\item{target_variable}{The target variable to predict.}

\item{output_file}{The file name to save the univariate regression results.}
}
\description{
This function performs univariate logistic regression on the training data and saves the results.
}
